Hitasoft Task/requirements.txt
Django==5.2
djangorestframework==3.15.1
djangorestframework-authtoken==1.3.0
pytz==2024.1
sqlparse==0.4.4
asgiref==3.8.1
